tataaig/requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.1
numpy==1.19.1
pandas==1.1.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
XlsxWriter==1.3.3
xlwt==1.3.0
